% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{Beta}
\alias{Beta}
\title{Beta distribution}
\usage{
Beta(shape1, shape2, mean, sd, ...)
}
\arguments{
\item{shape1, shape2}{Shape parameters of the beta distribution.}

\item{mean, sd}{Mean and standard deviation of the distribution, as an
alternative to \code{shape1}/\code{shape2}.}

\item{...}{Limits of the distribution, passed to \code{\link[=bound_dist]{bound_dist()}}.}
}
\value{
A \verb{<dist_spec>}.
}
\description{
A beta distribution as a \verb{<dist_spec>}, given either by its shape parameters
\code{shape1}/\code{shape2} or by its \code{mean}/\code{sd}. It is not discretised.
}
\examples{
Beta(shape1 = 2, shape2 = 5)
Beta(mean = 0.3, sd = 0.15)
}
\seealso{
\link{Distributions} for an overview and the other distributions.
}
