% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{convert_to_logmean}
\alias{convert_to_logmean}
\title{Convert mean and sd to log mean for a log normal distribution}
\usage{
convert_to_logmean(mean, sd)
}
\arguments{
\item{mean}{Numeric, mean of a distribution}

\item{sd}{Numeric, standard deviation of a distribution}
}
\value{
The log mean of a lognormal distribution
}
\description{
Convert from mean and standard deviation to the log mean of the
lognormal distribution. Useful for defining a \code{\link[=LogNormal]{LogNormal()}} distribution
from a mean and standard deviation on the natural scale.
}
\examples{

convert_to_logmean(2, 1)
}
