% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed.R
\name{Fixed}
\alias{Fixed}
\title{Fixed (point-mass) distribution}
\usage{
Fixed(value, ...)
}
\arguments{
\item{value}{Value of the fixed (delta) distribution.}

\item{...}{Limits of the distribution, passed to \code{\link[=bound_dist]{bound_dist()}}.}
}
\value{
A \verb{<dist_spec>}.
}
\description{
A fixed (delta) distribution as a \verb{<dist_spec>}, placing all of its mass on a
single \code{value}.
}
\examples{
Fixed(value = 3)
Fixed(value = 3.5)
}
\seealso{
\link{Distributions} for an overview and the other distributions.
}
