% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{Gamma}
\alias{Gamma}
\title{Gamma distribution}
\usage{
Gamma(shape, rate, scale, mean, sd, ...)
}
\arguments{
\item{shape, scale}{shape and scale parameters.  Must be positive,
    \code{scale} strictly.}

\item{rate}{an alternative way to specify the scale.}

\item{mean, sd}{Mean and standard deviation of the distribution, as an
alternative to \code{shape}/\code{rate}.}

\item{...}{Limits of the distribution, passed to \code{\link[=bound_dist]{bound_dist()}}.}
}
\value{
A \verb{<dist_spec>}.
}
\description{
A gamma distribution as a \verb{<dist_spec>}, given either by its natural
parameters \code{shape}/\code{rate} (or \code{shape}/\code{scale}) or by its \code{mean}/\code{sd}.
}
\examples{
Gamma(mean = 4, sd = 1)
Gamma(shape = 16, rate = 4)
Gamma(shape = Normal(16, 2), rate = Normal(4, 1))
}
\seealso{
\link{Distributions} for an overview and the other distributions.
}
