% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal.R
\name{LogNormal}
\alias{LogNormal}
\title{Lognormal distribution}
\usage{
LogNormal(meanlog, sdlog, mean, sd, ...)
}
\arguments{
\item{meanlog, sdlog}{mean and standard deviation of the distribution
    on the log scale with default values of \code{0} and \code{1} respectively.}

\item{mean, sd}{Mean and standard deviation of the distribution, as an
alternative to \code{meanlog}/\code{sdlog}.}

\item{...}{Limits of the distribution, passed to \code{\link[=bound_dist]{bound_dist()}}.}
}
\value{
A \verb{<dist_spec>}.
}
\description{
A lognormal distribution as a \verb{<dist_spec>}, given either by its natural
parameters \code{meanlog}/\code{sdlog} or by its \code{mean}/\code{sd}.
}
\examples{
LogNormal(mean = 4, sd = 1)
LogNormal(mean = 4, sd = 1, max = 10)
# Uncertain parameters must be given as the natural parameters
LogNormal(meanlog = Normal(1.5, 0.5), sdlog = 0.25, max = 10)
}
\seealso{
\link{Distributions} for an overview and the other distributions.
}
