% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_spec.R
\name{lower_bounds}
\alias{lower_bounds}
\title{Get the lower bounds of the parameters of a distribution}
\usage{
lower_bounds(x)
}
\arguments{
\item{x}{A \verb{<dist_spec>}.}
}
\value{
A numeric vector, the lower bounds.
}
\description{
This is used to avoid sampling parameter values that have no support.
}
\examples{
lower_bounds(LogNormal(meanlog = 0, sdlog = 1))
}
\keyword{internal}
