% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{natural_params}
\alias{natural_params}
\title{Get the names of the natural parameters of a distribution}
\usage{
natural_params(x)
}
\arguments{
\item{x}{A \verb{<dist_spec>}.}
}
\value{
A character vector, the natural parameters.
}
\description{
These are the natural (canonical) parameters of a distribution, such as
\code{shape} and \code{rate} for a gamma distribution or \code{meanlog} and \code{sdlog} for a
lognormal distribution. All other parameter representations (for example a
mean and standard deviation) are converted to these using
\code{\link[=convert_to_natural]{convert_to_natural()}}.
}
\examples{
natural_params(Gamma(shape = 1, rate = 1))
# a distribution type can also be given by name
natural_params("gamma")
}
\keyword{internal}
