% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_dist_spec}
\alias{new_dist_spec}
\title{Internal function for generating a \code{dist_spec} given parameters and a
distribution.}
\usage{
new_dist_spec(
  params,
  distribution,
  max = Inf,
  tail_cutoff = 0,
  cdf_cutoff = lifecycle::deprecated()
)
}
\arguments{
\item{params}{Parameters of the distribution (including \code{max})}

\item{distribution}{Character; the distribution type (e.g. \code{"gamma"},
\code{"lognormal"}, \code{"nonparametric"}).}

\item{max}{Numeric, maximum value of the distribution. The distribution will
be truncated at this value. Default: \code{Inf}, i.e. no maximum.}

\item{tail_cutoff}{Numeric; the desired CDF cutoff. Any part of the
cumulative distribution function beyond 1 minus the value of this argument is
removed. Default: \code{0}, i.e. use the full distribution.}

\item{cdf_cutoff}{Deprecated; use \code{tail_cutoff} instead.}
}
\value{
A \code{dist_spec} of the given specification.
}
\description{
This will convert all parameters to natural parameters before generating
a \code{dist_spec}. If they are uncertain the uncertainty is propagated to the
natural parameters with a first-order (delta-method) approximation (see
\code{\link[=convert_to_natural]{convert_to_natural()}}).
}
\examples{
new_dist_spec(
  params = list(mean = 2, sd = 1),
  distribution = "normal"
)
}
