% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{Normal}
\alias{Normal}
\title{Normal distribution}
\usage{
Normal(mean, sd, ...)
}
\arguments{
\item{mean, sd}{Mean and standard deviation of the distribution.}

\item{...}{Limits of the distribution, passed to \code{\link[=bound_dist]{bound_dist()}}.}
}
\value{
A \verb{<dist_spec>}.
}
\description{
A normal distribution as a \verb{<dist_spec>}, given by its \code{mean} and \code{sd}. Also
used to give an uncertain parameter of another distribution.
}
\examples{
Normal(mean = 4, sd = 1)
Normal(mean = 4, sd = 1, max = 10)
}
\seealso{
\link{Distributions} for an overview and the other distributions.
}
