% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{natural_params}
\alias{natural_params}
\title{Get the names of the natural parameters of a distribution}
\usage{
natural_params(x)
}
\arguments{
\item{x}{A \verb{<dist_spec>}.}
}
\value{
A character vector, the natural parameters.
}
\description{
These are the parameters used in the stan models. All other parameter
representations are converted to these using \code{\link[=convert_to_natural]{convert_to_natural()}} before
being passed to the stan models.
}
\examples{
natural_params(Gamma(shape = 1, rate = 1))
# a distribution type can also be given by name
natural_params("gamma")
}
\keyword{internal}
